+++ src/financial/history/trend_summary.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 from src.financial.history.trend_direction import (
     FinancialMomentum,
@@ -11,7 +12,7 @@ class MetricTrend:
     """Represents the trend for one financial metric."""
 
     direction: TrendDirection
-    change: float
+    change: Decimal
 
     def to_dict(self) -> dict:
         """Convert the metric trend to a dictionary."""
